LibVideo/VP9: Write blocks' primary inter predictions into framebuffer
Previously, we had an intermediate buffer to write the inter-predicted
block to, which would then be copied to the framebuffer. We can skip
the copy and write directly to the framebuffer by simply providing a
stride to the block prediction, and then the compound frame blending
can just read, blend and write back to the framebuffer.

This results in about a 7.6% speedup (~10.78s -> ~9.94s).