LibVideo/VP9: Use an enum for TransformSet, with a value for lossless
Previously, we were checking whether the frame was lossless for each
row and column in a transform, but by changing the transform sets to an
enum we can include that lossless mode as one of them without adding
any complexity to the transform selection in `inverse_transform_2d()`.

This has the added benefit of allowing us to drop the `BlockContext`
parameter to the inverse transform functions, which seems help with
their optimization a lot.