LibVideo/VP9: Skip transforming rows that contain zero coefficients
There will be many blocks that only include a few non-zero coefficients
in their residuals. We can avoid doing as much work for these by only
running the second transform's inverse for the rows that have non-zero
coefficients and skipping the rest.

In cases where we have no non-zero coefficients, we don't even have to
do anything for the second transform, and can skip straight to the
first transform's inverse.